+++ yodl/src/hashmap/hashmapremove.c
@@ -1,6 +1,6 @@
 #include "hashmap.ih"
 
-void hashmap_remove(register HashMap *map, unsigned idx)
+void hashmap_remove(register HashMap *map, size_t idx)
 {
     free(map->d_map[idx]);
     map->d_map[idx] = (HashItem *)~0;
