+++ yodl/src/parser/parserifcond.c
@@ -2,7 +2,7 @@
 
 void parser_if_cond(register Parser *pp,
             bool (*comparator)(Parser *pp, char **parlist),
-            char const *fun, unsigned nparlists)
+            char const *fun, size_t nparlists)
 {
     register char **parlist;
     int idx = 0;
