+++ yodl/src/stack/stackcontains.c
@@ -3,7 +3,7 @@
 bool stack_contains(Stack *sp, void const *value)
 {
     register void **vp = sp->d_value;
-    register unsigned idx;
+    register size_t idx;
 
     for (idx = sp->d_n; idx--; vp++)
     {
